+++ src/main/java/components/control/VSliderComponent.java
@@ -188,7 +188,7 @@ public class VSliderComponent extends ACtrlComponent {
             case KeyEvent.VK_ENTER:
                 fireEventAdjustmentBegin();
                 try {
-                    setValue(Float.parseFloat(keybBuffer));
+                    setValue(Double.parseDouble(keybBuffer));
                 } catch (java.lang.NumberFormatException ex) {
                 }
                 fireEventAdjustmentFinished();
